Still working
Think theres a threading issue. Will try moving the actionlistener out
of trombonescreen to see if i cant set up a simulated keyboard input to
get it to play as though someone was typing at the keys.

+++ trombone/Trombone.java
@@ -49,9 +49,9 @@ public class Trombone {
 
 	}
 	
-	public void playNote(Note n, double length) throws InterruptedException{
+	public void playNote(Note n, long length) throws InterruptedException{
 		this.playNote(n);
-		Thread.sleep((long) length);
+		Thread.sleep( length);
 		this.endNote(n);
 
 	}
